Moved the ClassVisitor interface into the Visitors package
Moved the ClassVisitor interface into the visitors package since all implementing classes are inside of the visitors package or the http sub-package.
